requirements.txt
streamlit
numpy
tensorflow
keras
Pillow
opencv-python